academy_monitor/requirements.txt
selenium==4.25.0
webdriver-manager==4.0.2
pillow==10.4.0
pystray==0.19.5
playsound==1.3.0
python-dotenv==1.0.1
requests==2.32.3